refactor: move project checkers to the checkers package
Align the project level rules with the checkers/rules split introduced in
#1821: rules modules keep only the Rule classes while the project checkers
live in robocop.linter.checkers.
